telaf: Fix the build issue using /opt/qct/ toolchain when building in Yocto
The script findtoolchain is used to get the toolchain path and pass
back to Makefile. The logic of the function is to try to find toolchain
in Yocto first, and then in /opt/qct. The issue is that the later finding
will overwrite the previous one. The fix is to return if the toolchain is
found in Yocto.

Change-Id: I8b7e9944888de1e95bfabd8af3ab9fb032a450b3
Signed-off-by: Yan Kong <xiankong@codeaurora.org>