TelAF: Support importing RW/RO cfg file to config tree
Import the .cfg file which was bundled as [w] attribute in
.cdef to config tree with the .cfg file name as the root
node, but if there is a same root node name already exist
in the tree file, the import action will stop.

And if it was bundled as [r] attribute, the framework will
keep everything under this root node same as that cfg file.

Change-Id: Ic323b88380e513073ff7a27e41b3599489eef0f0